tests/wpt/web-platform-tests/css/css-multicol/parsing/column-rule-style-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: getComputedValue().columnRuleStyle</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-rule-style">
<meta name="assert" content="column-rule-style computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("column-rule-style", "none");
test_computed_value("column-rule-style", "hidden");
test_computed_value("column-rule-style", "dotted");
test_computed_value("column-rule-style", "dashed");
test_computed_value("column-rule-style", "solid");
test_computed_value("column-rule-style", "double");
test_computed_value("column-rule-style", "groove");
test_computed_value("column-rule-style", "ridge");
test_computed_value("column-rule-style", "inset");
</script>
</body>
</html>
